+++ kitty/gl.h
@@ -66,6 +66,7 @@ ArrayInformation program_uniform_array(int program, const char *name);
 ssize_t create_vao(void);
 size_t add_buffer_to_vao(ssize_t vao_idx, GLenum usage);
 void add_attribute_to_vao(ssize_t vao_idx, int location, GLint size, GLenum data_type, GLsizei stride, void *offset, GLuint divisor);
+void set_vao_attribute(ssize_t vao_idx, size_t buffer_idx, int location, GLint size, GLenum data_type, GLsizei stride, void *offset, GLuint divisor);
 ssize_t alloc_vao_buffer(ssize_t vao_idx, GLsizeiptr size, size_t bufnum, GLenum usage);
 void* alloc_and_map_vao_buffer(ssize_t vao_idx, GLsizeiptr size, size_t bufnum, bool frequently_updated);
 void unmap_vao_buffer(ssize_t vao_idx, size_t bufnum);